(PC-23923) refactor(tutorial): move AgeInformationTutorial in tutorial feature
